work-reporting-backend/WorkReportingRestApi/requirements.txt
Django
djangorestframework
psycopg2
django-cors-headers
djangorestframework-jwt